lesson14: Валидация, FormRequest

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
@@ -22,6 +23,7 @@ class CategoryController extends Controller
     public function index():View|Factory|Application
     {
         $categories = Category::get();
+
         return view('auth.categories.index', compact('categories'));
     }
 
@@ -42,12 +44,14 @@ class CategoryController extends Controller
      *
      * @return RedirectResponse
      */
-    public function store(Request $request):RedirectResponse
+    public function store(CategoryRequest $request):RedirectResponse
     {
-        $path = $request->file('image')->store('categories');
-        //image - название поля html верстке в input у кнопки "Загрузить", categories - папка для загрузки картинок
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {//если в запросе есть картинка, то мы добавляем сохранение
+            $params['image'] = $request->file('image')->store('categories');
+            //image - название поля html верстке в input у кнопки "Загрузить", categories - папка для загрузки картинок
+        }
         Category::create($params);
         return redirect()->route('categories.index');
     }
@@ -84,12 +88,14 @@ class CategoryController extends Controller
      *
      * @return RedirectResponse
      */
-    public function update(Request $request, Category $category):RedirectResponse
+    public function update(CategoryRequest $request, Category $category):RedirectResponse
     {
-        Storage::delete('image');
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) { //проверка на существование картинки
+            Storage::delete('image');
+            $params['image']  = $request->file('image')->store('categories');
+        }
         $category->update($params);
         return redirect()->route('categories.index');
     }
@@ -104,6 +110,7 @@ class CategoryController extends Controller
     public function destroy(Category $category):RedirectResponse
     {
         $category->delete();
+
         return redirect()->route('categories.index');
     }
 }
